Use PlotGroups explicitly
If this commit is applied then plotJuggler groups will be used instead
of series so that if a topic is keyed then it is easy to add its
instance to the group.
Stored all groups pointer into an unordered map so that I can find back
the group based on topic names.

Note that I have used default separator "/", might be an issue if it
changes.

Issue: [https://github.com/eProsima/fastdds-visualizer-plugin/issues/56]
Signed-off-by: Manuel Valch <manuelValch@proton.me>